+++ app/__init__.py
@@ -2,8 +2,8 @@ from flask import Flask
 
 app = Flask(__name__)
 
-app.config['VERSION'] = '0.15.0'
-app.config['VERSION_DATE'] = '20 March 2026'
+app.config['VERSION'] = '0.16.0'
+app.config['VERSION_DATE'] = '21 March 2026'
 
 @app.context_processor
 def inject_version():